requirements.txt
flask>=2.2
pandas>=2.0
openpyxl>=3.0
python-dateutil>=2.8
gunicorn>=20.1
pyinstaller>=5.8
pywebview>=3.6
